pyproject.toml
[tool.poetry]
package-mode = false

[project]
name = "freelancehunt-automizer"
version = "0.1.0"
description = ""
authors = [
    {name = "Vitaliy Fedytnyk",email = "vitalikfedytnyk@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "psycopg-binary (>=3.2.9,<4.0.0)",
    "sqlalchemy (>=2.0.43,<3.0.0)",
    "pydantic (>=2.11.7,<3.0.0)",
    "openai (>=1.100.2,<2.0.0)",
    "selenium (>=4.35.0,<5.0.0)",
    "alembic (>=1.16.4,<2.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "beautifulsoup4 (>=4.13.4,<5.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
